Refactor the implementation and bug fixing
- Extract the logic about error handling out of the classpath view.
- Remove some redundent effects.
- Fix dropdown css bugs.
- Fix the bug that activeProjectIndex not currently used in handlers.
- Fix the bug when handling the library paths.
