+++ src/extension.ts
@@ -22,9 +22,7 @@ import { initialize as initUtils } from "./utils";
 import { KEY_SHOW_WHEN_USING_JAVA } from "./utils/globalState";
 import { scheduleAction } from "./utils/scheduler";
 import { showWelcomeWebview, WelcomeViewSerializer } from "./welcome";
-import { activateCopilotInspection } from "./copilot/inspect";
 import { ProjectSettingsViewSerializer } from "./project-settings/projectSettingsView";
-import { isLlmApiReady } from "./copilot/utils";
 
 let cleanJavaWorkspaceIndicator: string;
 let activatedTimestamp: number;
