+++ src/project-settings/assets/classpath/style.scss
@@ -35,6 +35,7 @@
 
 .setting-section-dropdown {
   width: 320px;
+  z-index: 10;
 
   .dropdown-description {
     position: sticky;
