+++ src/project-settings/assets/mainpage/features/commonSlice.ts
@@ -3,7 +3,7 @@
 
 import { createSlice } from "@reduxjs/toolkit";
 import { ProjectType } from "../../../../utils/webview";
-import { ProjectState, SectionId } from "../../../types";
+import { SectionId } from "../../../types";
 import _ from "lodash";
 
 export const commonSlice = createSlice({
@@ -12,32 +12,39 @@ export const commonSlice = createSlice({
         ui: {
             activeSection: SectionId.Classpath,
             activeProjectIndex: 0,
+            exception: undefined,
+            loadingState: false,
         },
         data: {
             projects: [],
             projectType: [] as ProjectType[],
-            projectState: [] as ProjectState[],
         }
     },
     reducers: {
         updateActiveSection: (state, action) => {
             state.ui.activeSection = action.payload;
         },
+        catchException: (state, action) => {
+          state.ui.exception = action.payload;
+        },
+        updateLoadingState: (state, action) => {
+          state.ui.loadingState = action.payload;
+        },
         listProjects: (state, action) => {
             state.data.projects = action.payload;
             const projectNum = state.data.projects.length;
             state.data.projectType = Array(projectNum).fill("");
-            state.data.projectState = Array(projectNum).fill(ProjectState.Unloaded);
         },
         activeProjectChange: (state, action) => {
             state.ui.activeProjectIndex = action.payload;
-            if (state.data.projectType[state.ui.activeProjectIndex] === ProjectType.Maven
+            // If the active project is not a Maven project, the Maven section is invisible,
+            // switch to the Classpath section.
+            if (state.data.projectType[state.ui.activeProjectIndex] !== ProjectType.Maven
                     && state.ui.activeSection === SectionId.Maven) {
                         state.ui.activeSection = SectionId.Classpath;
             }
         },
         setProjectType: (state, action) => {
-            state.data.projectState[state.ui.activeProjectIndex] = ProjectState.Loaded;
             state.data.projectType[state.ui.activeProjectIndex] = action.payload.projectType;
         },
     },
@@ -45,6 +52,8 @@ export const commonSlice = createSlice({
 
 export const {
     updateActiveSection,
+    catchException,
+    updateLoadingState,
     listProjects,
     activeProjectChange,
     setProjectType,
